move accessibility setup to accessibility module
